Update BubbleSort.cpp
int to double

+++ BubbleSort.cpp
@@ -1,6 +1,6 @@
 #include "BubbleSort.h"
 
-void BubbleSort::sort(int array[], int size)
+void BubbleSort::sort(double array[], int size)
 {
 	for (int i = 0; i < size; i++) {
 		for (int j = 0; j < size - 1; j++) {
